pyproject.toml
[project]
name = "teamsoverflow-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.1.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pyright>=1.1.399",
    "ruff>=0.11.4",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
